refactor(repair-tickets): decompose RepairTicketsService into Lifecycle/Query/Warranty
Behavior-preserving (10 $transaction, 0 JE). Facade keeps 14-method surface + 6-dep ctor
+ delegates; internally constructs 3 sub-services (specs + module + defect-exchange consumer
untouched). 10 $tx moved WHOLE; markReplaced kept facade-delegated with EXACT defaulted-tx
signature (defect-exchange external-tx caller intact); returnToCustomer cross-module
createDraftForRepair (SHOP ExpenseDocument / CUSTOMER OtherIncome) tx-passthrough preserved.
tsc 0, repair-tickets tests green, lint 0.
